refactor(browser): keep protocol backend neutral
